Refactor when query methods into a trait
